+++ frontend/src/components/common/HorizontalBarChart/HorizontalBarChart.js
