+++ frontend/src/components/common/index.js
@@ -20,6 +20,7 @@ import MapChartSvg from './MapChartSvg/MapChartSvg';
 import MapChartStandard from './MapChartStandard/MapChartStandard';
 import LottieAnimation from './LottieAnimation/LottieAnimation';
 import { PublicRoute, PrivateRoute } from './Route/Route';
+import HorizontalBarChart from './HorizontalBarChart/HorizontalBarChart';
 
 export {
   Avatar,
@@ -45,4 +46,5 @@ export {
   LottieAnimation,
   PublicRoute,
   PrivateRoute,
+  HorizontalBarChart,
 };
