+++ frontend/src/components/index.js
@@ -11,7 +11,6 @@ import CommentItem from './CommentItem/CommentItem';
 import RegionalStateChart from './RegionalStateChart/RegionalStateChart';
 import RegionalStateTable from './RegionalStateTable/RegionalStateTable';
 import ReviewFilterList from './ReviewFilterList/ReviewFilterList';
-import HorizontalBarChart from './HorizontalBarChart/HorizontalBarChart';
 import ImageModal from './ImageModal/ImageModal';
 import ReviewImage from './ReviewImage/ReviewImage';
 
@@ -29,7 +28,6 @@ export {
   RegionalStateChart,
   RegionalStateTable,
   ReviewFilterList,
-  HorizontalBarChart,
   ImageModal,
   ReviewImage,
 };
